Unite update interceptor tests
