+++ packages/test/src/test-integration-update-interceptors.ts
@@ -3,8 +3,6 @@ import * as wf from '@temporalio/workflow';
 import { Next, UpdateInput, WorkflowInboundCallsInterceptor, WorkflowInterceptors } from '@temporalio/workflow';
 import { helpers, makeTestFunction } from './helpers-integration';
 
-const update = wf.defineUpdate<string[], [string]>('update');
-
 const test = makeTestFunction({
   workflowsPath: __filename,
   workflowInterceptorModules: [__filename],
@@ -23,24 +21,27 @@ const test = makeTestFunction({
   },
 });
 
-export async function workflowWithUpdates(): Promise<string[]> {
-  const state: string[] = [];
-  const updateHandler = async (arg: string): Promise<string[]> => {
-    state.push(arg);
-    if (arg === 'fail-update') {
-      throw new wf.ApplicationFailure(`Deliberate ApplicationFailure in handler`);
+const update = wf.defineUpdate<string, [string]>('update');
+
+export async function workflowWithUpdates(): Promise<void> {
+  const updateHandler = async (arg: string): Promise<string> => arg;
+  const validator = (arg: string): void => {
+    if (arg === 'bad-arg') {
+      throw new Error('Validation failed');
     }
-    return state;
   };
-  wf.setHandler(update, updateHandler);
-  await wf.condition(() => state.includes('done'));
-  state.push('$');
-  return state;
+  wf.setHandler(update, updateHandler, { validator });
+  await wf.condition(() => false); // Ensure the update is handled if it is dispatched in a second WFT.
 }
 
 class UpdateInboundCallsInterceptor implements WorkflowInboundCallsInterceptor {
   async handleUpdate(input: UpdateInput, next: Next<UpdateInboundCallsInterceptor, 'handleUpdate'>): Promise<unknown> {
     return await next({ ...input, args: [input.args[0] + '-inboundIntercepted', ...input.args.slice(1)] });
+  }
+  validateUpdate(input: UpdateInput, next: Next<UpdateInboundCallsInterceptor, 'validateUpdate'>): void {
+    const [arg] = input.args as string[];
+    const args = arg.startsWith('validation-interceptor-will-make-me-invalid') ? ['bad-arg'] : [arg];
+    next({ ...input, args });
   }
 }
 
@@ -55,7 +56,7 @@ test('Update client and inbound interceptors work for executeUpdate', async (t) 
     const wfHandle = await startWorkflow(workflowWithUpdates);
 
     const updateResult = await wfHandle.executeUpdate(update, { args: ['1'] });
-    t.deepEqual(updateResult, ['1-clientIntercepted-inboundIntercepted']);
+    t.deepEqual(updateResult, '1-clientIntercepted-inboundIntercepted');
   });
 });
 
@@ -67,6 +68,20 @@ test('Update client and inbound interceptors work for startUpdate', async (t) =>
 
     const updateHandle = await wfHandle.startUpdate(update, { args: ['1'] });
     const updateResult = await updateHandle.result();
-    t.deepEqual(updateResult, ['1-clientIntercepted-inboundIntercepted']);
+    t.deepEqual(updateResult, '1-clientIntercepted-inboundIntercepted');
+  });
+});
+
+test('Update validation interceptor works', async (t) => {
+  const { createWorker, startWorkflow, assertWorkflowUpdateFailed } = helpers(t);
+  const worker = await createWorker();
+  await worker.runUntil(async () => {
+    const wfHandle = await startWorkflow(workflowWithUpdates);
+    await assertWorkflowUpdateFailed(
+      wfHandle.executeUpdate(update, { args: ['validation-interceptor-will-make-me-invalid'] }),
+      wf.ApplicationFailure,
+      'Validation failed'
+    );
+    t.pass();
   });
 });
